+++ src/docs/changelog.md
@@ -8,6 +8,7 @@ Changelog                           {#changelog}
 
 - #407 more renaming iShaderMaterial to iShader and related
 - removed widget example since Mica covers all you need to know about widgets in Igor
+- #374 using float32 for mouse position to reduce conversions
 
 0.44.0
 ------